made database_utils to contain database info and metods, also added restraint, that the same url adress isnt added to db
