+++ aplikuj_blueprint.py
@@ -10,49 +10,7 @@ from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from bs4 import BeautifulSoup
 from datetime import datetime, timedelta
-
-def get_database_connection():
-    server = 'joboffers.database.windows.net'
-    database = 'JobOffersDB'
-    username = 'Nobody'
-    password = 'Karwala1'
-    driver = '{ODBC Driver 17 for SQL Server}'
-    connection_string = f'DRIVER={driver};SERVER=tcp:{server};PORT=1433;DATABASE={database};UID={username};PWD={password};Encrypt=yes;TrustServerCertificate=no;Connection Timeout=30;'
-    return pyodbc.connect(connection_string)
-
-def get_or_create_website(cursor, website_name, website_url):
-    cursor.execute("SELECT id FROM api_websites WHERE Website_url = ?", (website_url,))
-    result = cursor.fetchone()
-    if result:
-        return result[0]
-    else:
-        cursor.execute("INSERT INTO api_websites (Website_name, Website_url) OUTPUT INSERTED.id VALUES (?, ?)", (website_name, website_url))
-        return cursor.fetchone()[0]
-
-def get_or_create_category(cursor, category_name):
-    cursor.execute("SELECT id FROM api_categories WHERE Category_name = ?", (category_name,))
-    result = cursor.fetchone()
-    if result:
-        return result[0]
-    else:
-        cursor.execute("INSERT INTO api_categories (Category_name) OUTPUT INSERTED.id VALUES (?)", (category_name,))
-        return cursor.fetchone()[0]
-
-def insert_offer_data(offer_data):
-    conn = get_database_connection()
-    cursor = conn.cursor()
-
-    website_id = get_or_create_website(cursor, offer_data['Website_name'], offer_data['Website'])
-    category_id = get_or_create_category(cursor, offer_data['Category'])
-
-    insert_query = """INSERT INTO api_joboffers (Position, Website_id, Category_id, Firm, Earnings, Location, Date, Job_type, Link) 
-                      VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?);"""
-    cursor.execute(insert_query, (
-        offer_data['Position'], website_id, category_id, offer_data['Firm'], offer_data['Earnings'], 
-        offer_data['Location'], offer_data['Date'], offer_data['Job_type'], offer_data['Link']))
-    conn.commit()
-    cursor.close()
-    conn.close()
+from database_utils import insert_offer_data
 
 def transform_date(publication_date):
     months = {
@@ -168,7 +126,6 @@ def scrapp(site_url, category_name, category_path):
 
 
 aplikuj_blueprint = func.Blueprint()
-
 
 @aplikuj_blueprint.timer_trigger(schedule="0 01 00 * * *", arg_name="myTimer", run_on_startup=False,
               use_monitor=False)
