+++ manpower_blueprint.py
@@ -128,7 +128,7 @@ def scrapp(site_url, category_name, category_path):
 
 manpower_blueprint = func.Blueprint()
 
-@manpower_blueprint.timer_trigger(schedule="0 01 00 * * *", arg_name="myTimer", run_on_startup=True,
+@manpower_blueprint.timer_trigger(schedule="0 01 00 * * *", arg_name="myTimer", run_on_startup=False,
               use_monitor=False) 
 def manpower_timer_trigger(myTimer: func.TimerRequest) -> None:
     if myTimer.past_due:
